Changed recordMatch() to be on stub instead of responseConfig
In the process, simplified the design considerably. The code used
to save state through out-of-process protocol proxying to record
the actual response. This is unneeded complexity for what is simply
a debug tool; now it only saves the proxy configuration in such
scenarios. I also chose not to implement it with the file backed
repository, although the code now supports adding it in the future.
https://github.com/bbyars/mountebank/issues/279