Converted _behaviors to array
The behaviors object executes the behaviors in a non-docuemnted
order but allows lists for shellTransform, copy, lookup. The new
format accepts a list and executes them in the order provided, which
(for example) allows you to do 2 lookups on either side of a decorate,
giving the user full control over the pipeline.
https://github.com/bbyars/mountebank/issues/521